move compute_nbins to spatial utils
