+++ spiketools/tests/spatial/test_occupancy.py
@@ -10,17 +10,6 @@ from spiketools.spatial.occupancy import *
 ###################################################################################################
 ###################################################################################################
 
-def test_compute_nbins():
-
-    # check 1d case
-    out1 = compute_nbins([5, 5])
-    assert out1 == 25
-
-    # check 2d case
-    out2 = compute_nbins([5])
-    out3 = compute_nbins(5)
-    assert out2 == out3 == 5
-
 def test_compute_bin_edges():
 
     # check 1d case
